Decompose extended example
